test: start MCP deadlines after transport setup
Move the native MCP timeout case into the focused evidence suite and connect the stdio transport before registering the 1.5-second turn deadline. This keeps the production timeout unchanged while preventing slow hosted-runner process startup from expiring the token before the tested invocation begins.
