+++ tests/native-tool-evidence.test.ts
@@ -6,6 +6,10 @@ import { join } from "node:path";
 import { compileChatGptWebPrompt } from "../src/adapters/chatgpt-web/prompt";
 import { readTextFileCommand } from "../src/adapters/chatgpt-web/native-command";
 import { callTurnBroker, TurnBroker } from "../src/adapters/chatgpt-web/turn-broker";
+import {
+  ChatGptExternalTurnProgress,
+  chatGptExternalToolCallsAreInFlight,
+} from "../src/adapters/chatgpt-web/turn-progress";
 import { CHATGPT_WEB_MODEL_ID } from "../src/adapters/chatgpt-web/model";
 import { defaultBrokerEndpoint } from "../src/config";
 import type { ChatGptTurnEnvironment } from "../src/adapters/chatgpt-web/environment";
@@ -14,6 +18,7 @@ import {
   chatGptMcpInvocationTimeout,
 } from "../src/adapters/chatgpt-web/mcp-server";
 import type { CodexParsedRequest } from "../src/types";
+import { toolResult } from "./chatgpt-harness-fixture";
 
 const evidenceContract = "Describe failed local actions using only observable tool evidence. If no native result was returned, state only that the action did not execute; never infer or name an unreported cause.";
 const safeDiscoveryContract = "__codex_tool_search__:<capability query>";
@@ -194,6 +199,80 @@ test("an aborted MCP request revokes only its turn binding and leaves the server
     await client.close().catch(() => {});
     broker.revoke(abandonedToken);
     broker.revoke(replacementToken);
+    await broker.close();
+  }
+}, 10_000);
+
+test("a native tool deadline starts after MCP transport setup and retires its browser boundary", async () => {
+  const socketPath = brokerEndpoint(`cgw-native-timeout-${process.pid}-${Date.now()}`);
+  const broker = TurnBroker.forSocket(socketPath);
+  const environment: ChatGptTurnEnvironment = {
+    cwd: process.cwd(),
+    roots: [process.cwd()],
+    writableRoots: [process.cwd()],
+    sandboxPolicy: { type: "dangerFullAccess" },
+    tools: [{ name: "exec_command", description: "Run a command", parameters: { type: "object" } }],
+  };
+  const client = new Client({ name: "native-timeout-test", version: "1.0.0" });
+  const tokens: string[] = [];
+
+  try {
+    await client.connect(new StdioClientTransport({
+      command: process.execPath,
+      args: ["src/cli.ts", "mcp", "--broker-socket", socketPath],
+      cwd: process.cwd(),
+      stderr: "pipe",
+    }));
+    const timedOutToken = await broker.register(environment, 1_500, "timeout-turn");
+    const replacementToken = await broker.register(environment, undefined, "replacement-turn");
+    tokens.push(timedOutToken, replacementToken);
+
+    const timedOut = client.callTool({
+      name: "codex_exec",
+      arguments: { turn_token: timedOutToken, cmd: "slow external MCP call" },
+    });
+    const [request] = await broker.nextToolBatch(timedOutToken);
+    expect(request).toMatchObject({ wireName: "exec_command" });
+    const externalProgress = new ChatGptExternalTurnProgress();
+    const revision = externalProgress.recordToolBatch(1, 1_000);
+    const pendingBatch = externalProgress.waitForToolBatchObservation(revision).then(() => request);
+    const batchOutcome = pendingBatch.then(
+      value => ({ type: "value" as const, value }),
+      error => ({ type: "error" as const, error: error instanceof Error ? error : new Error(String(error)) }),
+    );
+    const retirement = broker.waitForRetirement(timedOutToken).then(() => {
+      externalProgress.retire(new Error("MCP invocation retired its turn binding"));
+    });
+
+    const timeoutResult = await timedOut;
+    expect(timeoutResult.isError).toBe(true);
+    expect(timeoutResult.structuredContent).toMatchObject({
+      code: "codex_tool_timeout",
+      tool: "exec_command",
+      retryable: false,
+    });
+    expect(JSON.stringify(timeoutResult.content)).toContain("did not complete before the MCP transport deadline");
+    await retirement;
+    expect(externalProgress.snapshot().activeToolCalls).toBe(0);
+    expect(chatGptExternalToolCallsAreInFlight(externalProgress.snapshot())).toBeFalse();
+    const settledBatch = await batchOutcome;
+    expect(settledBatch.type).toBe("error");
+    if (settledBatch.type !== "error") throw new Error("retired tool batch crossed its browser boundary");
+    expect(settledBatch.error.message).toContain("retired its turn binding");
+    expect(() => externalProgress.recordToolResult()).toThrow("retired its turn binding");
+    await expect(callTurnBroker(socketPath, { method: "claim", token: timedOutToken }))
+      .rejects.toThrow("already finished");
+    expect(() => broker.completeTool(timedOutToken, request!.callId, toolResult({ output: "late" })))
+      .toThrow("turn token is invalid or expired");
+
+    const inventory = await client.callTool({
+      name: "codex_tool_inventory",
+      arguments: { turn_token: replacementToken, query: "exec_command", include_schema: false },
+    });
+    expect(inventory.structuredContent).toMatchObject({ total: 1, tools: [{ wire_name: "exec_command" }] });
+  } finally {
+    await client.close().catch(() => {});
+    for (const token of tokens) broker.revoke(token);
     await broker.close();
   }
 }, 10_000);
